Separated the evidence endpoints from config
